# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cabinet/overview": {
      "get": {
        "description": "Возвращает только данные владельца активной сессии; не содержит subscription URL, хешей или VPN-ключей.",
        "operationId": "CabinetController_overview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "subscription",
                    "devices"
                  ],
                  "properties": {
                    "subscription": {
                      "nullable": true,
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "status",
                        "planName",
                        "deviceLimit",
                        "startsAt",
                        "expiresAt"
                      ],
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "PENDING",
                            "ACTIVE",
                            "EXPIRED",
                            "CANCELLED"
                          ]
                        },
                        "planName": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128
                        },
                        "deviceLimit": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "startsAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        }
                      }
                    },
                    "devices": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "id",
                          "displayName",
                          "platform",
                          "status",
                          "createdAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "displayName": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 128,
                            "nullable": true
                          },
                          "platform": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 32,
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "ACTIVE",
                              "REVOKED"
                            ]
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Сессия отсутствует, истекла или отозвана"
          }
        },
        "summary": "Получить безопасную сводку кабинета текущего пользователя",
        "tags": [
          "cabinet"
        ]
      }
    },
    "/cabinet/devices": {
      "post": {
        "description": "Проверяет активную подписку, лимит устройств и Origin кабинета. Полный subscription URL возвращается только в этом ответе и не сохраняется в API.",
        "operationId": "CabinetController_issueDevice",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "description": "UUID generated by the cabinet for safe request retry",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "origin",
            "in": "header",
            "description": "Trusted cabinet origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "id",
                    "displayName",
                    "platform",
                    "status",
                    "createdAt",
                    "subscriptionUrl"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "displayName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128,
                      "nullable": true
                    },
                    "platform": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 32,
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "ACTIVE"
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "subscriptionUrl": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректное тело запроса"
          },
          "401": {
            "description": "Сессия отсутствует, истекла или отозвана"
          },
          "403": {
            "description": "Недоверенный Origin кабинета"
          },
          "409": {
            "description": "Нет активной подписки или исчерпан лимит"
          },
          "503": {
            "description": "Выдача не настроена или нет доступной HEALTHY-ноды"
          }
        },
        "summary": "Выпустить устройство текущего пользователя",
        "tags": [
          "cabinet"
        ]
      }
    },
    "/cabinet/devices/{deviceId}/revoke": {
      "post": {
        "description": "Идемпотентно отзывает только принадлежащее текущему пользователю устройство и ставит обновление desired state для связанных нод.",
        "operationId": "CabinetController_revokeDevice",
        "parameters": [
          {
            "name": "origin",
            "in": "header",
            "description": "Trusted cabinet origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Устройство отозвано или уже отозвано"
          },
          "400": {
            "description": "Некорректный идентификатор устройства"
          },
          "401": {
            "description": "Сессия отсутствует, истекла или отозвана"
          },
          "403": {
            "description": "Недоверенный Origin кабинета"
          },
          "404": {
            "description": "Устройство пользователя не найдено"
          }
        },
        "summary": "Отозвать устройство текущего пользователя",
        "tags": [
          "cabinet"
        ]
      }
    },
    "/sub/{token}": {
      "get": {
        "description": "Bearer token устройства проверяется сервером. Renderer выдаёт только подтверждённые VLESS/TCP/TLS/HAPP маршруты; отсутствие готового маршрута означает временную инфраструктурную недоступность.",
        "operationId": "SubscriptionFeedController_feed",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Opaque device subscription token",
            "schema": {
              "type": "string",
              "minLength": 43,
              "maxLength": 43
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Текстовый subscription feed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "maxLength": 16384
                }
              }
            }
          },
          "401": {
            "description": "Неверный, отозванный или истёкший token"
          },
          "429": {
            "description": "Too many requests"
          },
          "503": {
            "description": "Entitlement действует, но готовый VPN-маршрут недоступен"
          }
        },
        "summary": "Получить subscription feed устройства",
        "tags": [
          "subscription-feed"
        ]
      }
    },
...
    "/prototype/subscription/{token}": {
      "get": {
        "description": "Только для закрытого локального прототипа; endpoint выключен по умолчанию. Содержимое fixture задаётся только через некоммитимую локальную переменную окружения.",
        "operationId": "SubscriptionPrototypeController_feed",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Opaque local-only prototype token",
            "schema": {
              "minLength": 32,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Текстовый fixture subscription-ответа",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "# VPNPlatform local subscription prototype\\n"
                }
              }
            }
          },
          "401": {
            "description": "Неверный или отсутствующий токен"
          },
          "404": {
            "description": "Локальный прототип выключен"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "summary": "Получить локальный fixture subscription-ответа",
        "tags": [
          "subscription-prototype"
        ]
      }
    }
  },
  "info": {
    "title": "VPNPlatform API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http"
      }
    },
...
  }
}