# apps/api/src/cabinet/cabinet.controller.ts
import {
  BadRequestException,
  Body,
  Controller,
  Get,
  Header,
  Headers,
  HttpCode,
  Inject,
  Param,
  Post,
  UnauthorizedException,
  UseGuards,
} from '@nestjs/common';
import {
  ApiOkResponse,
  ApiOperation,
  ApiCreatedResponse,
  ApiBadRequestResponse,
  ApiConflictResponse,
  ApiForbiddenResponse,
  ApiHeader,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiParam,
  ApiServiceUnavailableResponse,
  ApiTags,
  ApiUnauthorizedResponse,
} from '@nestjs/swagger';
import {
  cabinetDeviceIdempotencyKeySchema,
  cabinetDeviceIdSchema,
  createCabinetDeviceRequestSchema,
  type CabinetOverview,
  type IssuedCabinetDevice,
} from '@vpn-platform/contracts';

import { AuthSessionService } from '../auth/auth-session.service';
import { TrustedOriginGuard } from '../auth/trusted-origin.guard';
import { CabinetService } from './cabinet.service';
import { CabinetDeviceService } from './cabinet-device.service';

@ApiTags('cabinet')
@Controller('cabinet')
export class CabinetController {
  constructor(
    @Inject(AuthSessionService) private readonly sessions: AuthSessionService,
    @Inject(CabinetService) private readonly cabinet: CabinetService,
    @Inject(CabinetDeviceService)
    private readonly devices: CabinetDeviceService,
  ) {}

  @Get('overview')
  @ApiOperation({
    summary: 'Получить безопасную сводку кабинета текущего пользователя',
    description:
      'Возвращает только данные владельца активной сессии; не содержит subscription URL, хешей или VPN-ключей.',
  })
  @ApiOkResponse({ schema: cabinetOverviewOpenApiSchema() })
  @ApiUnauthorizedResponse({
    description: 'Сессия отсутствует, истекла или отозвана',
  })
  async overview(
    @Headers('cookie') cookieHeader: string | undefined,
  ): Promise<CabinetOverview> {
    const session = await this.sessions.currentSessionFromCookie(cookieHeader);
    if (!session) {
      throw new UnauthorizedException('Session is invalid');
    }

    return this.cabinet.overview(session.user.id);
  }

  @Post('devices')
  @Header('Cache-Control', 'no-store')
  @UseGuards(TrustedOriginGuard)
  @ApiOperation({
    summary: 'Выпустить устройство текущего пользователя',
    description:
      'Проверяет активную подписку, лимит устройств и Origin кабинета. Полный subscription URL возвращается только в этом ответе и не сохраняется в API.',
  })
  @ApiHeader({
    name: 'origin',
    required: true,
    description: 'Trusted cabinet origin',
  })
  @ApiHeader({
    name: 'idempotency-key',
    required: true,
    description: 'UUID generated by the cabinet for safe request retry',
    schema: { type: 'string', format: 'uuid' },
  })
  @ApiCreatedResponse({ schema: issuedCabinetDeviceOpenApiSchema() })
  @ApiBadRequestResponse({ description: 'Некорректное тело запроса' })
  @ApiForbiddenResponse({ description: 'Недоверенный Origin кабинета' })
  @ApiConflictResponse({
    description: 'Нет активной подписки или исчерпан лимит',
  })
  @ApiServiceUnavailableResponse({
    description: 'Выдача не настроена или нет доступной HEALTHY-ноды',
  })
  @ApiUnauthorizedResponse({
    description: 'Сессия отсутствует, истекла или отозвана',
  })
  async issueDevice(
    @Body() body: unknown,
    @Headers('cookie') cookieHeader: string | undefined,
    @Headers('idempotency-key') idempotencyKey: string | undefined,
  ): Promise<IssuedCabinetDevice> {
    const input = createCabinetDeviceRequestSchema.safeParse(body);
    if (!input.success) {
      throw new BadRequestException('Device request is invalid');
    }
    const parsedIdempotencyKey =
      cabinetDeviceIdempotencyKeySchema.safeParse(idempotencyKey);
    if (!parsedIdempotencyKey.success) {
      throw new BadRequestException('Device idempotency key is invalid');
    }
    const session = await this.sessions.currentSessionFromCookie(cookieHeader);
    if (!session) {
      throw new UnauthorizedException('Session is invalid');
    }

    return this.devices.issue(
      session.user.id,
      parsedIdempotencyKey.data,
      input.data,
    );
  }

  @Post('devices/:deviceId/revoke')
  @HttpCode(204)
  @UseGuards(TrustedOriginGuard)
  @ApiOperation({
    summary: 'Отозвать устройство текущего пользователя',
    description:
      'Идемпотентно отзывает только принадлежащее текущему пользователю устройство и ставит обновление desired state для связанных нод.',
  })
  @ApiParam({ name: 'deviceId', schema: { type: 'string', format: 'uuid' } })
  @ApiHeader({
    name: 'origin',
    required: true,
    description: 'Trusted cabinet origin',
  })
  @ApiNoContentResponse({ description: 'Устройство отозвано или уже отозвано' })
  @ApiBadRequestResponse({
    description: 'Некорректный идентификатор устройства',
  })
  @ApiForbiddenResponse({ description: 'Недоверенный Origin кабинета' })
  @ApiNotFoundResponse({ description: 'Устройство пользователя не найдено' })
  @ApiUnauthorizedResponse({
    description: 'Сессия отсутствует, истекла или отозвана',
  })
  async revokeDevice(
    @Param('deviceId') deviceId: string,
    @Headers('cookie') cookieHeader: string | undefined,
  ): Promise<void> {
    const parsedDeviceId = cabinetDeviceIdSchema.safeParse(deviceId);
    if (!parsedDeviceId.success) {
      throw new BadRequestException('Device id is invalid');
    }
    const session = await this.sessions.currentSessionFromCookie(cookieHeader);
    if (!session) {
      throw new UnauthorizedException('Session is invalid');
    }
    await this.devices.revoke(session.user.id, parsedDeviceId.data);
  }
}

function issuedCabinetDeviceOpenApiSchema() {
  return {
    type: 'object',
    additionalProperties: false,
    required: [
      'id',
      'displayName',
      'platform',
      'status',
      'createdAt',
      'subscriptionUrl',
    ],
    properties: {
      id: { type: 'string', format: 'uuid' },
      displayName: {
        type: 'string',
        minLength: 1,
        maxLength: 128,
        nullable: true,
      },
      platform: { type: 'string', minLength: 1, maxLength: 32, nullable: true },
      status: { type: 'string', enum: ['ACTIVE'] },
      createdAt: { type: 'string', format: 'date-time' },
      subscriptionUrl: { type: 'string', format: 'uri' },
    },
  };
}

function cabinetOverviewOpenApiSchema() {
  return {
    type: 'object',
    additionalProperties: false,
    required: ['subscription', 'devices'],
    properties: {
      subscription: {
        nullable: true,
        type: 'object',
        additionalProperties: false,
        required: [
          'status',
          'planName',
          'deviceLimit',
          'startsAt',
          'expiresAt',
        ],
        properties: {
          status: {
            type: 'string',
            enum: ['PENDING', 'ACTIVE', 'EXPIRED', 'CANCELLED'],
          },
          planName: { type: 'string', minLength: 1, maxLength: 128 },
          deviceLimit: { type: 'integer', minimum: 1 },
          startsAt: { type: 'string', format: 'date-time', nullable: true },
          expiresAt: { type: 'string', format: 'date-time', nullable: true },
        },
      },
      devices: {
        type: 'array',
        items: {
          type: 'object',
          additionalProperties: false,
          required: ['id', 'displayName', 'platform', 'status', 'createdAt'],
          properties: {
            id: { type: 'string', format: 'uuid' },
            displayName: {
              type: 'string',
              minLength: 1,
              maxLength: 128,
              nullable: true,
            },
            platform: {
              type: 'string',
              minLength: 1,
              maxLength: 32,
              nullable: true,
            },
            status: { type: 'string', enum: ['ACTIVE', 'REVOKED'] },
            createdAt: { type: 'string', format: 'date-time' },
          },
        },
      },
    },
  };
}